create ScrimBoxMiddleArea.jsx
