+++ client/src/components/scrim_components/ScrimSection.jsx
@@ -26,6 +26,7 @@ import ShareIcon from '@material-ui/icons/Share';
 import SettingsIcon from '@material-ui/icons/Settings';
 import UploadPostGameImage from './UploadPostGameImage';
 import pluralize from 'pluralize';
+import ScrimBoxMiddleArea from './ScrimBoxMiddleArea';
 
 const compareDates = (scrim) => {
   let currentTime = new Date().getTime();
@@ -170,9 +171,6 @@ export default function ScrimSection({ scrim, isInDetail }) {
       }
     }
   };
-
-  const teamOneDifference = 5 - teamOne.length;
-  const teamTwoDifference = 5 - teamTwo.length;
 
   // the "Scrim Box"
   return (
@@ -342,192 +340,16 @@ export default function ScrimSection({ scrim, isInDetail }) {
             gameStarted={gameStarted === scrim._id}
           />
 
-          <div className={classes.teamsVersusSeparator}>
-            <div
-              className="lobby__info-box"
-              style={{
-                background: `rgba(255, 255, 255,${
-                  imageUploaded === scrim._id
-                    ? '0.8'
-                    : gameStarted === scrim._id
-                    ? '0.7'
-                    : '0.5'
-                })`,
-                padding: '10px',
-                borderRadius: '4px',
-              }}>
-              {gameStarted !== scrim._id && (
-                <h2 className="text-black">Game starting in...</h2>
-              )}
-
-              <CountdownTimer
-                gameStarted={gameStarted}
-                setGameStarted={setGameStarted}
-                scrim={scrim}
-              />
-
-              {gameStarted === scrim._id &&
-                (scrim.teamOne.length === 5 && scrim.teamTwo.length === 5 ? (
-                  <>
-                    {!gameEnded && (
-                      <>
-                        {/* show lobby name and pswd only to players in lobby or admins */}
-                        {playerEntered ||
-                        casterEntered ||
-                        process.env.REACT_APP_ADMIN_KEY ===
-                          currentUser?.adminKey ? (
-                          <>
-                            <h2 className="text-black">
-                              Lobby host / captain: {scrim.lobbyHost?.name}
-                            </h2>
-                            <h3 className="text-black">
-                              please make the lobby name: <br />"
-                              {scrim.lobbyName}"
-                            </h3>
-                            <h3 className="text-black">
-                              with the password: {scrim.lobbyPassword}
-                            </h3>
-                          </>
-                        ) : (
-                          <></>
-                        )}
-                      </>
-                    )}
-
-                    {/* WHO WON ? BUTTONS */}
-                    {/* show buttons if is admin or is lobby captain */}
-                    {/* don't show if game has ended */}
-                    {(scrim.lobbyHost.email === currentUser?.email ||
-                      currentUser?.adminKey ===
-                        process.env.REACT_APP_ADMIN_KEY) &&
-                      !gameEnded && (
-                        // WHO WON BUTTONS
-                        <Grid
-                          item
-                          container
-                          alignItems="center"
-                          direction="row"
-                          spacing={2}>
-                          <Grid item>
-                            <h3 className="text-black">Who won?</h3>
-                          </Grid>
-
-                          {['Team 1 (Blue Side)', 'Team 2 (Red Side)'].map(
-                            (teamTitle, idx) => (
-                              <Grid item key={idx}>
-                                <Button
-                                  style={{
-                                    backgroundColor: idx === 0 ? 'blue' : 'red',
-                                    color: '#fff',
-                                  }}
-                                  variant="contained"
-                                  onClick={async () => {
-                                    // set team won for scrim
-                                    let yes =
-                                      window.confirm(`Are you sure ${teamTitle} won this game? \n 
-                                  You cannot reverse this.
-                                  `);
-
-                                    if (!yes) return;
-
-                                    const dataSending = {
-                                      teamWon: teamTitle,
-                                    };
-                                    const updatedScrim = await updateScrim(
-                                      scrim._id,
-                                      dataSending
-                                    );
-                                    if (updatedScrim) {
-                                      getNewScrimsData();
-                                    }
-                                  }}>
-                                  {teamTitle}
-                                </Button>
-                              </Grid>
-                            )
-                          )}
-                        </Grid>
-                      )}
-
-                    {/*  allow image upload if both teams are filled and 
-                    the current user is the host or creator of scrim or an admin.
-                  */}
-                    {/* POST GAME IMAGE SECTION */}
-                    {(scrim.lobbyHost?._id === currentUser?._id ||
-                      currentUser?.adminKey ===
-                        process.env.REACT_APP_ADMIN_KEY) && (
-                      <>
-                        <Box marginTop={2} />
-
-                        <UploadPostGameImage
-                          isUploaded={imageUploaded === scrim._id}
-                          scrim={scrim}
-                        />
-                      </>
-                    )}
-
-                    {imageUploaded === scrim._id && (
-                      <Grid
-                        item
-                        container
-                        direction="row"
-                        justifyContent="space-between"
-                        alignItems="center"
-                        spacing={2}>
-                        <Grid item>
-                          <h3 className="text-black">
-                            Post-game image uploaded!
-                          </h3>
-                        </Grid>
-                        <Grid item>
-                          <Button
-                            variant="contained"
-                            color="primary"
-                            component="a"
-                            href={scrim.postGameImage?.location}
-                            rel="noreferrer"
-                            target="_blank">
-                            View Image
-                          </Button>
-                        </Grid>
-                      </Grid>
-                    )}
-                  </>
-                ) : (
-                  <>
-                    <h2 className="text-black">
-                      Not enough players:&nbsp;
-                      {`${teamOne.length + teamTwo.length}/10`}
-                    </h2>
-                    <h5 className="text-black">
-                      Please get
-                      {/* if teamOne still needs players show this else don't show */}
-                      {teamOneDifference > 0 ? (
-                        <>
-                          &nbsp;{teamOneDifference}&nbsp;
-                          {/* spell check singular and plural with pluralize */}
-                          {pluralize('players', teamOneDifference)} in Team
-                          1&nbsp;
-                          <br />
-                        </>
-                      ) : null}
-                      {/* if teamTwo needs players, show this text. */}
-                      {teamTwoDifference > 0 ? (
-                        <>
-                          {/* if team one needs players, show 'and', else don't show 'and' */}
-                          {teamOneDifference > 0 ? 'and' : ''}&nbsp;
-                          {teamTwoDifference}
-                          &nbsp;{pluralize('players', teamTwoDifference)} in
-                          Team 2
-                          <br />
-                        </>
-                      ) : null}
-                      to unlock lobby name and password
-                    </h5>
-                  </>
-                ))}
-            </div>
-          </div>
+          {/* the middle box that contains the countdown timer and other details. */}
+          <ScrimBoxMiddleArea
+            imageUploaded={imageUploaded === scrim._id}
+            scrim={scrim}
+            gameStarted={gameStarted === scrim._id}
+            setGameStarted={setGameStarted}
+            gameEnded={gameEnded}
+            playerEntered={playerEntered}
+            casterEntered={casterEntered}
+          />
 
           {/* teamTwo */}
           <ScrimTeamList
